Refactor: 금액 타입을 int에서 BigDecimal로 변경

+++ src/main/java/kr/hhplus/be/server/application/order/OrderResult.java
@@ -3,6 +3,8 @@ package kr.hhplus.be.server.application.order;
 import kr.hhplus.be.server.domain.payment.PaymentInfo;
 import lombok.Builder;
 
+import java.math.BigDecimal;
+
 public class OrderResult {
 
     public static Payment of(PaymentInfo.Payment payment) {
@@ -15,7 +17,7 @@ public class OrderResult {
     public record Payment(
             Long orderId,
             Long userId,
-            int paymentAmount
+            BigDecimal paymentAmount
     ) {
         @Builder
         public Payment {}
