+++ src/main/java/kr/hhplus/be/server/domain/payment/PaymentService.java
@@ -4,6 +4,8 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.math.BigDecimal;
+
 @Service
 @RequiredArgsConstructor
 public class PaymentService {
@@ -11,7 +13,7 @@ public class PaymentService {
     private final PaymentRepository paymentRepository;
 
     @Transactional
-    public PaymentInfo.Payment payment(Long orderId, int totalAmount, int discountAmount) {
+    public PaymentInfo.Payment payment(Long orderId, BigDecimal totalAmount, BigDecimal discountAmount) {
         Payment payment = Payment.create(orderId, totalAmount, discountAmount);
         Payment savedPayment = paymentRepository.save(payment);
         return PaymentInfo.of(savedPayment);
